% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_dominate.R
\name{f_dominate}
\alias{f_dominate}
\title{Determining cover balls}
\usage{
f_dominate(x_main, x_other, proportion = 1, p_of = 0)
}
\arguments{
\item{x_main}{asd}

\item{x_other}{asd}

\item{proportion}{asd}

\item{p_of}{asd}
}
\value{
asd
 \item{results}{asd}
}
\description{
Determining cover balls
}
\details{
asd
}
\examples{
rnorm(1)

}
\author{
Fatih Saglam, saglamf89@gmail.com
}
