% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DatRel.R
\name{DatRel}
\alias{DatRel}
\title{Data Relocation for Resampled Data using Pure and Proper Class Cover Catch Digraph}
\usage{
DatRel(x, y, x_syn, proportion = 1, p_of = 0, class_pos = NULL)
}
\arguments{
\item{x}{feature matrix or dataframe.}

\item{y}{class factor variable.}

\item{x_syn}{asd}

\item{proportion}{proportion of covered samples. A real number between \eqn{(0,1]}.
1 by default. Smaller numbers results in less dominant samples.}

\item{p_of}{asd}

\item{class_pos}{asd}
}
\value{
an list which includes:
 \item{x_new}{dominant sample indexes.}
 \item{y_new}{dominant samples from feature matrix, x}
 \item{x_syn}{Radiuses of the circle for dominant samples}
 \item{i_dominant}{class names}
 \item{x_pos_dominant}{number of classes}
 \item{radii_pos_dominant}{proportions each class covered}
}
\description{
\code{DatRel} relocates resampled data using Pure and Proper Class Cover Catch Digraph
}
\details{
asd
}
\examples{

rnorm(1)

}
\references{
Priebe, C. E., DeVinney, J. G., & Marchette, D. J. (2001). On the distribution
of the domination number for random class cover catch digraphs. Statistics &
Probability Letters, 55(3), 239-246.

Marchette, C. E. P. D. J., & Socolinsky, J. G. D. D. A. (2003).
 Classiﬁcation Using Class Cover Catch Digraphs. Journal of Classiﬁcation, 20, 3-23.

Manukyan, A., & Ceyhan, E. (2016). Classification of imbalanced data with a
geometric digraph family. The Journal of Machine Learning Research, 17(1), 6504-6543.
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
