% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_dominate.R
\name{f_dominate}
\alias{f_dominate}
\title{Determining cover balls}
\usage{
f_dominate(x_main, x_other, proportion = 1, p_of = 0)
}
\arguments{
\item{x_main}{Target class samples.}

\item{x_other}{Non-target class samples.}

\item{proportion}{proportion of covered samples. A real number between \eqn{(0,1]}.
1 by default. Smaller numbers results in less dominant samples.}

\item{p_of}{roportion to increase cover radius. A real number between
\eqn{(0,\infty)}. Default is 0. Higher values tolerate other classes more.}
}
\value{
a list object with following:
 \item{i_dominant}{dominant sample indexes}
 \item{dist_main2other}{distance matrix of target class samples to non-target
class samples}
 \item{dist_main2main}{distance matrix of target class samples to target
class samples}
}
\description{
Determining cover balls
}
\details{
To be used in \code{DatRel}.
}
\author{
Fatih Saglam, saglamf89@gmail.com
}
