% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_relocate.R
\name{f_relocate}
\alias{f_relocate}
\title{Relocation function}
\usage{
f_relocate(x_pos_dominant, x_syn, radii_pos_dominant, p_of = 0)
}
\arguments{
\item{x_pos_dominant}{asd}

\item{x_syn}{asd}

\item{radii_pos_dominant}{asd}

\item{p_of}{asd}
}
\value{
asd
 \item{x_syn_relocated}{relocated samples.}
}
\description{
Relocation function
}
\details{
asd
}
\examples{
rnorm(1)

}
\author{
Fatih Saglam, saglamf89@gmail.com
}
